+++ RELEASE NOTES.md
@@ -1,3 +1,9 @@
+Version 1.1
+
+- Added crossfade transitions
+- Added addition responder chain and view hierarchy methods
+- Renamed some hierarchy methods for better clarity
+
 Version 1.0
 
 - Initial release.